Delete the hale-runtime interpreter crate — retire slice 2
Slice 1 (#41) repointed `hale run` onto the codegen backend, so
the tree-walking interpreter has no remaining caller. This deletes
the crate (~10,400 lines): the `eval`/`value`/`bus`/`env`/
`builtins` evaluator and its 10 parity test files, plus the
workspace member and the hale-cli dependency.

Coverage: the interpreter's parity suite is superseded by the
compiled-corpus ASan oracle and the per-form codegen tests
(form_{vec,hashmap,ring_buffer}_codegen, recpool_codegen,
violate_build, violate_child_state_read, the stdlib_* suite). The
one parity test without a direct codegen twin —
method_call_from_free_fn (the 3a self-binding regression: a free
fn calling a locus method must bind the receiver so `self.X`
resolves, including through a nested method call) — is ported to
crates/hale-codegen/tests/method_call_from_free_fn.rs and passes.

Spec/docs follow the impl in the same commit: the three "###
Interpreter parity" subsections in runtime.md are removed; the
"both interpreter and codegen" time-clock notes collapse to the
single codegen path; the `hale run` cross-seed-import notes
(projects.md, semantics.md) now describe the codegen path
(single-file imports resolve; only the ad-hoc directory form
skips cross-seed renames). One user-visible reduction is recorded:
the interpreter-only `err.<capture_name>` convenience on
ClosureViolation is gone — captured state is read through the
child handle in `on_failure(c, err)` (semantics.md, design-
rationale.md, types.md, memory.md). The compiler-dev brief drops
the interpreter from the pipeline + crate map.
