+++ crates/hale-cli/Cargo.toml
@@ -13,7 +13,6 @@ path = "src/main.rs"
 [dependencies]
 hale-syntax = { path = "../hale-syntax" }
 hale-types = { path = "../hale-types" }
-hale-runtime = { path = "../hale-runtime" }
 hale-codegen = { path = "../hale-codegen" }
 serde = { version = "1", features = ["derive"] }
 toml = "0.8"
