--- crates/hale-runtime/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "hale-runtime"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-description = "Hale toolchain runtime — interpreter (v0) + region/scheduler/bus (Phase 2). Manages the lotus substrate."
-
-[dependencies]
-hale-syntax = { path = "../hale-syntax" }
-hale-types = { path = "../hale-types" }
-libc = "0.2"
