--- crates/hale-runtime/tests/form_vec_interpreter.rs
@@ -1,333 +1,0 @@
-//! v1.x-FORM-1 PR7: end-to-end `@form(vec)` execution under
-//! the interpreter. Each test parses + typechecks + runs a
-//! complete program. Codegen-side parity lands in a future
-//! FORM-2 PR (PR5/6) and gets its own test file.
-
-use hale_runtime::run_program;
-
-fn run(src: &str) -> i32 {
-    let program = hale_syntax::parse_source(src)
-        .map_err(|d| {
-            d.iter()
-                .map(|x| x.render(src))
-                .collect::<Vec<_>>()
-                .join("\n")
-        })
-        .expect("parse");
-    run_program(&program).expect("run")
-}
-
-fn run_expect_error(src: &str) -> String {
-    let program = hale_syntax::parse_source(src)
-        .map_err(|d| {
-            d.iter()
-                .map(|x| x.render(src))
-                .collect::<Vec<_>>()
-                .join("\n")
-        })
-        .expect("parse");
-    match run_program(&program) {
-        Ok(_) => panic!("expected program to exit with an error, got ok"),
-        Err(s) => s,
-    }
-}
-
-#[test]
-fn push_get_round_trip() {
-    let src = r#"
-        @form(vec)
-        locus ItemListL {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = ItemListL { };
-            l.push(42);
-            let head = l.get(0) or raise;
-            println(head);
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn len_and_is_empty_track_pushes() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = L { };
-            if !l.is_empty() { println("FAIL: should start empty"); }
-            l.push(1);
-            l.push(2);
-            l.push(3);
-            if l.len() != 3 { println("FAIL: len should be 3"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn pop_returns_last_element_lifo() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = L { };
-            l.push(10);
-            l.push(20);
-            l.push(30);
-            let a = l.pop() or raise;
-            let b = l.pop() or raise;
-            if a != 30 { println("FAIL: first pop"); }
-            if b != 20 { println("FAIL: second pop"); }
-            if l.len() != 1 { println("FAIL: len after pops"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn get_out_of_bounds_substitute_uses_fallback() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = L { };
-            l.push(7);
-            let v = l.get(99) or -1;
-            if v != -1 { println("FAIL: fallback not used"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn pop_empty_substitute_uses_fallback() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = L { };
-            let v = l.pop() or -42;
-            if v != -42 { println("FAIL: empty-pop fallback not used"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn get_out_of_bounds_or_raise_bubbles() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn main() {
-            let l = L { };
-            let v = l.get(99) or raise;
-            println(v);
-        }
-    "#;
-    // The unaddressed `raise` past the top-level fn surfaces as
-    // a runtime error in the interpreter (no on_failure to
-    // catch). Just verify the program errors out rather than
-    // succeeding.
-    let _err = run_expect_error(src);
-}
-
-#[test]
-fn err_binding_available_on_substitute_rhs() {
-    let src = r#"
-        @form(vec)
-        locus L {
-            capacity { heap items of Int; }
-        }
-        fn fallback(e: IndexError) -> Int {
-            return e.index + e.len;
-        }
-        fn main() {
-            let l = L { };
-            l.push(1);
-            l.push(2);
-            // Out-of-bounds at index 5; len=2; fallback should
-            // see 5 + 2 = 7.
-            let v = l.get(5) or fallback(err);
-            if v != 7 { println("FAIL: err binding payload"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn fail_stmt_exits_via_error_path() {
-    let src = r#"
-        type E { code: Int; }
-        fn pick(x: Int) -> Int fallible(E) {
-            if x < 0 {
-                fail E { code: 1 };
-            }
-            return x * 2;
-        }
-        fn main() {
-            let ok_v = pick(5) or raise;
-            let bad_v = pick(-3) or -99;
-            if ok_v != 10 { println("FAIL: success path"); }
-            if bad_v != -99 { println("FAIL: error fallback"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn set_overwrites_existing_index() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of Int; } }
-        fn main() {
-            let l = L { };
-            l.push(10);
-            l.push(20);
-            l.push(30);
-            l.set(1, 99) or raise;
-            let v = l.get(1) or raise;
-            if v != 99 { println("FAIL: set"); }
-            if (l.get(0) or raise) != 10 { println("FAIL: a"); }
-            if (l.get(2) or raise) != 30 { println("FAIL: c"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn set_out_of_bounds_substitute_uses_fallback() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of Int; } }
-        fn noop(_e: IndexError) { }
-        fn main() {
-            let l = L { };
-            l.push(1);
-            l.set(99, 0) or noop(err);
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn vec_of_struct_cells() {
-    let src = r#"
-        type Pair { x: Int; y: Int; }
-        @form(vec)
-        locus PairsL {
-            capacity { heap items of Pair; }
-        }
-        fn main() {
-            let l = PairsL { };
-            l.push(Pair { x: 1, y: 2 });
-            l.push(Pair { x: 3, y: 4 });
-            let first = l.get(0) or raise;
-            if first.x != 1 { println("FAIL: x"); }
-            if first.y != 2 { println("FAIL: y"); }
-            println("ok");
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-/// Sort family — interpreter parity with the codegen path. The
-/// program exits with `std::process::exit(1)` if any ordering
-/// invariant fails, so `run() == 0` actually tests behaviour.
-#[test]
-fn sort_int_ascending_interp() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of Int; } }
-        fn main() {
-            let l = L { };
-            l.push(3); l.push(1); l.push(4); l.push(1); l.push(5);
-            l.sort();
-            let mut i = 1;
-            while i < l.len() {
-                let prev = l.get(i - 1) or raise;
-                let cur = l.get(i) or raise;
-                if cur < prev { std::process::exit(1); }
-                i = i + 1;
-            }
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn sort_string_ascending_interp() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of String; } }
-        fn main() {
-            let l = L { };
-            l.push("delta"); l.push("alpha"); l.push("charlie");
-            l.sort();
-            let a = l.get(0) or raise;
-            let b = l.get(1) or raise;
-            let c = l.get(2) or raise;
-            if a != "alpha" { std::process::exit(1); }
-            if b != "charlie" { std::process::exit(1); }
-            if c != "delta" { std::process::exit(1); }
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn sort_by_custom_cmp_interp() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of Int; } }
-        fn gt(a: Int, b: Int) -> Bool { return a > b; }
-        fn main() {
-            let l = L { };
-            l.push(2); l.push(5); l.push(1); l.push(4);
-            l.sort_by(gt);
-            let a = l.get(0) or raise;
-            let b = l.get(3) or raise;
-            if a != 5 { std::process::exit(1); }
-            if b != 1 { std::process::exit(1); }
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
-
-#[test]
-fn sort_desc_by_interp() {
-    let src = r#"
-        @form(vec)
-        locus L { capacity { heap items of Int; } }
-        fn lt(a: Int, b: Int) -> Bool { return a < b; }
-        fn main() {
-            let l = L { };
-            l.push(2); l.push(5); l.push(1); l.push(4);
-            l.sort_desc_by(lt);
-            let a = l.get(0) or raise;
-            let b = l.get(3) or raise;
-            if a != 5 { std::process::exit(1); }
-            if b != 1 { std::process::exit(1); }
-        }
-    "#;
-    assert_eq!(run(src), 0);
-}
