+++ crates/hale-types/src/lib.rs
@@ -31,11 +31,11 @@ pub mod working_set;
 
 /// m94: subject wildcard matching used by the type checker
 /// (publish-side authorization for computed subjects) and
-/// mirrored at runtime by `hale-runtime::bus::subject_match`
-/// and the C runtime's `lotus_subject_match`. v0 supports a
+/// mirrored at runtime by the C runtime's `lotus_subject_match`.
+/// v0 supports a
 /// trailing `**` that matches *zero or more* remaining
 /// dot-separated segments — `log.app.**` matches the root
-/// `log.app` AND any descendant. All three implementations
+/// `log.app` AND any descendant. Both implementations
 /// must agree.
 pub fn wildcard_match(pattern: &str, subject: &str) -> bool {
     if let Some(prefix) = pattern.strip_suffix("**") {
